@@ -7,7 +7,7 @@ namespace WoWCombatLogParser.Common.Utility;
 
 public static class Conversion
 {
-    private static readonly Regex isNumber = new Regex(@"^([0-9]+|0x[0-9a-f]+)$", RegexOptions.Compiled);
+    private static readonly Regex isNumber = new(@"^([0-9]+|0x[0-9a-f]+)$", RegexOptions.Compiled);
     private static readonly Dictionary<Type, Func<string, object>> _convertableTypes = new()
     {
         { typeof(WowGuid), value => new WowGuid(value) },
@@ -18,10 +18,20 @@ public static class Conversion
         { typeof(string), value => value.Replace("\"", "") },
         { typeof(UnitFlag), value => new UnitFlag(Convert.ToUInt32(value, 16)) },
     };
+    private static readonly Dictionary<Type, object> _defaultValues = new()
+    {
+        { typeof(WowGuid), WowGuid.Empty },
+        { typeof(DateTime), DateTime.MinValue },
+        { typeof(decimal), 0 },
+        { typeof(int), 0 },
+        { typeof(bool), false },
+        { typeof(string), string.Empty },
+        { typeof(UnitFlag), new UnitFlag() },
+    };
 
     public static object GetValue(string value, Type type)
     {
-        if (value.In("", "nil")) return default;
+        if (value.In("", "nil")) return _defaultValues[type];
 
         if (_convertableTypes.ContainsKey(type))
         {
@@ -37,6 +47,7 @@ public static class Conversion
     }
 
     public static T GetValue<T>(string value) =>  (T)GetValue(value, typeof(T));
+    public static T GetValue<T>(ICombatLogDataField value) => (T)GetValue(value, typeof(T));
     public static object GetValue(ICombatLogDataField value, Type type) => GetValue(value.ToString(), type);
     private static int ConvertToInt(string value) => Convert.ToInt32(value, value.StartsWith("0x") ? 16 : 10);
     private static object ConvertToEnum(string value, Type type) =>  isNumber.IsMatch(value) ? Enum.ToObject(type, ConvertToInt(value)) : EnumExtensions.FromDescription(value, type);
