@@ -1,32 +1,37 @@
 ﻿using System;
 using System.Diagnostics;
+using System.Xml.Linq;
 using WoWCombatLogParser.Common.Events;
 
 namespace WoWCombatLogParser.Common.Models;
 
 [DebuggerDisplay("{Id} {UnitName} {Flags} {RaidFlags}")]
+public class Actor : Unit
+{
+    public UnitFlag Flags { get; set; }
+    public RaidFlag RaidFlags { get; set; }
+}
+
 public class Unit : CombatLogEventComponent, IKey
 {
     private string _name;
     private string _server;
-
-    [Key(4)]
     public WowGuid Id { get; set; }
     public string UnitName
     {
-        get => $"{_name}{(string.IsNullOrWhiteSpace(_server) ? "" : $" - {_server}")}";
+        get => $"{_name}{(string.IsNullOrWhiteSpace(_server) ? "" : $"-{_server}")}";
         set
         {
-            var values = value?.Split(new[] { " - " }, StringSplitOptions.RemoveEmptyEntries);
+            var values = value?.Split(new[] { "-" }, StringSplitOptions.RemoveEmptyEntries);
             for (int i = 0; i < values?.Length; i++)
             {
                 switch (i)
                 {
                     case 0:
-                        _name = values[i];
+                        _name = values[i].Trim();
                         break;
                     case 1:
-                        _server = values[i];
+                        _server = values[i].Trim();
                         break;
                     default:
                         break;
@@ -34,8 +39,6 @@ public class Unit : CombatLogEventComponent, IKey
             }
         }
     }
-    public UnitFlag Flags { get; set; }
-    public RaidFlag RaidFlags { get; set; }
 
     [NonData]
     public string Name => _name;
