@@ -52,8 +52,8 @@ and some duplicate text in the string", "\r\n", 39, 41)]
             ms.LastIndexOf(target).Should().Be(expectedIndex);
     }
 
-    private List<long> expectedStreamPositions = new List<long> { 8238838, 45390211, 46176415, 85248769, 91475221, 108816175, 139942758, 141175852, 158074360, 166786792, 183822650, 195043065, 213229974, 220542975, 239317560, 256811855, 282797912, 287168160, 295013244 };
-    private List<int> expectedStringPositions = new List<int> { 8234486, 45362942, 46148845, 85194857, 91418153, 108748573, 139855516, 141087938, 157976424, 166683070, 183709022, 194922976, 213100654, 220410125, 239175042, 256660782, 282633375, 287001464, 294842850 };
+    private List<long> expectedStreamPositions = [8238838, 45390211, 46176415, 85248769, 91475221, 108816175, 139942758, 141175852, 158074360, 166786792, 183822650, 195043065, 213229974, 220542975, 239317560, 256811855, 282797912, 287168160, 295013244];
+    private List<int> expectedStringPositions = [8234486, 45362942, 46148845, 85194857, 91418153, 108748573, 139855516, 141087938, 157976424, 166683070, 183709022, 194922976, 213100654, 220410125, 239175042, 256660782, 282633375, 287001464, 294842850];
 
     [Fact]
     public void Test_StreamExtensions_IndexOf_On_File()
