@@ -1,24 +1,41 @@
 ﻿using System;
 using System.IO;
+using WoWCombatLogParser.IO;
 using WoWCombatLogParser.Utility;
 
 namespace WoWCombatLogParser;
 
 public interface ICombatLogParser
 {
-    IParserContext ParserContext { get; set; }
+    IParserContext? ParserContext { get; set; }
     IEnumerable<Segment> GetSegments(string filename);
-    CombatLogEvent ParseLine(string text);
+    CombatLogEvent? ParseLine(string text);
     IEnumerable<CombatLogEvent> ParseSegment(Segment segment);
+    Dictionary<string, List<ParserError>> Errors { get; }
 }
 
 public class CombatLogParser : ICombatLogParser
 {
-    public IParserContext ParserContext { get; set; }
-    
-    public CombatLogEvent ParseLine(string line)
+    public IParserContext? ParserContext { get; set; }
+    public Dictionary<string, List<ParserError>> Errors { get; } = [];
+
+    public CombatLogEvent? ParseLine(string line)
     {
-        return ParserContext.EventGenerator.GetCombatLogEvent<CombatLogEvent>(line);
+        var data = CombatLogFieldReader.ReadFields(line);
+        try
+        {
+            return ParserContext?.EventGenerator.GetCombatLogEvent<CombatLogEvent>(data);
+        }
+        catch (Exception exception)
+        {
+            var eventType = exception is CombatLogParserException parserException ? parserException.EventType : data.EventType;
+            if (!Errors.TryGetValue(eventType, out var errors))
+            {
+                Errors[eventType] = errors = [];
+            }
+            errors.Add(new(exception, line));
+            return null;
+        }
     }
 
     public IEnumerable<Segment> GetSegments(string filename)
@@ -50,6 +67,16 @@ public class CombatLogParser : ICombatLogParser
     public IEnumerable<CombatLogEvent> ParseSegment(Segment segment)
     {
         foreach (var line in segment.Content)
-            yield return ParseLine(line);
+        {
+            var combatLogEvent = ParseLine(line);
+            if (combatLogEvent is { })
+                yield return combatLogEvent;
+        }
     }
+}
+
+public struct ParserError(Exception exception, string line)
+{
+    public Exception Exception = exception;
+    public string Raw = line;
 }
