@@ -4,7 +4,7 @@ namespace WoWCombatLogParser.Models;
 
 internal interface IParsable
 {
-    Task Parse(IList<ICombatLogDataField> data);
+    Task Parse(List<ICombatLogDataField> data);
     Task Parse(string line);
     IParserContext ParserContext { get; set; }
 }
