@@ -7,23 +7,26 @@ public interface IParserContext
 {
     ICombatLogParser CombatLogParser { get; set; }
     IEventGenerator EventGenerator { get; set; }
+    ICombatLogEventMapper CombatLogEventMapper { get; set; }
 }
 
 public class ParserContext : IParserContext
 {
     private ICombatLogParser combatLogParser;
     private IEventGenerator eventGenerator;
+    private ICombatLogEventMapper combatLogEventMapper;
     private IMapper Mapper;
 
-    public ParserContext() : this(new CombatLogParser(), new EventGenerator())
+    public ParserContext() : this(new CombatLogParser(), new EventGenerator(), new CombatLogEventMapper())
     {
         Mapper = InitializeMapper();
     }
 
-    public ParserContext(ICombatLogParser combatLogParser, IEventGenerator eventGenerator)
+    public ParserContext(ICombatLogParser combatLogParser, IEventGenerator eventGenerator, ICombatLogEventMapper combatLogEventMapper)
     {
         CombatLogParser = combatLogParser;
         EventGenerator = eventGenerator;
+        CombatLogEventMapper = combatLogEventMapper;
     }
 
     public ICombatLogParser CombatLogParser
@@ -44,6 +47,12 @@ public class ParserContext : IParserContext
             eventGenerator = value;
             eventGenerator.ParserContext = this;
         }
+    }
+
+    public ICombatLogEventMapper CombatLogEventMapper
+    {
+        get => combatLogEventMapper;
+        set => combatLogEventMapper = value;
     }
 
     private static IMapper InitializeMapper()
